requirements.txt
transformers
torch
google-generativeai
python-dotenv